Fix incorrect data types in pixel ufuncs
All of the pixel ufuncs use Healpix_Base2, which has 64 bit pixel
indices. However, it declares arguments as longs, which are not
necessarily a 64 bit data type (on armhf, they are 32 bit).

Fixing this allows us to manipulate 64 bit pixel indices even on
32 bit machines.

+++ healpy/test/test_pixelfunc.py
@@ -162,7 +162,7 @@ class TestPixelFunc(unittest.TestCase):
             numPix = nside2npix(nside)
             numRings = 4 * nside - 1  # Expected number of rings
             for nest in (True, False):
-                pix = np.arange(numPix)
+                pix = np.arange(numPix, dtype=np.int64)
                 ring = pix2ring(nside, pix, nest=nest)
                 self.assertTrue(pix.shape == ring.shape)
                 self.assertTrue(len(set(ring)) == numRings)
